万洽AI客服2.0，全新ai自动接待，知识库等功能发布

+++ OneChat-modules/OneChat-im/src/main/java/chat/aikf/im/tio/utils/KfCacheRelUtils.java
@@ -2,16 +2,33 @@ package chat.aikf.im.tio.utils;
 
 
 import chat.aikf.common.core.config.OneChatConfig;
+import chat.aikf.common.core.constant.Constants;
 import chat.aikf.common.core.constant.OneChatCacheKeyConstants;
+import chat.aikf.common.core.constant.SecurityConstants;
+import chat.aikf.common.core.domain.R;
 import chat.aikf.common.core.utils.SpringUtils;
 import chat.aikf.common.redis.service.RedisService;
-import chat.aikf.im.tio.model.GuestIdentityMsgDto;
+import chat.aikf.im.api.constant.OneChatMsgTypes;
+import chat.aikf.im.api.domain.dto.GuestIdentityMsgDto;
+import chat.aikf.im.tio.model.GuestReplyMsgDto;
+import chat.aikf.im.tio.model.VisitorSessionKey;
+import chat.aikf.im.tio.starter.OneChatImStarter;
+import chat.aikf.ops.api.RemoteKfRuleService;
+import chat.aikf.ops.api.RemoteKfVisitorService;
+import chat.aikf.ops.api.constant.OneChatVisitorSate;
+import chat.aikf.ops.api.domain.OneChatKfRule;
+import chat.aikf.ops.api.domain.OneChatKfRuleScope;
 import chat.aikf.ops.api.domain.OneChatKfVisitorMsg;
 import chat.aikf.ops.api.domain.OneChatkfVisitor;
 import chat.aikf.ops.api.utils.RuleFfServingUtils;
+import cn.hutool.json.JSONUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
+import org.tio.core.Tio;
+import org.tio.websocket.common.WsResponse;
 
+import java.util.Date;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
@@ -31,6 +48,24 @@ public class KfCacheRelUtils {
 
     @Autowired
     private RuleFfServingUtils ruleFfServingUtils;
+
+
+    @Autowired
+    private RemoteKfRuleService remoteKfRuleService;
+
+
+
+    @Autowired
+    @Lazy
+    private OneChatImStarter oneChatImStarter;
+
+
+
+    @Autowired
+    private RemoteKfVisitorService remoteKfVisitorService;
+
+
+
 
 
     /**
@@ -57,7 +92,7 @@ public class KfCacheRelUtils {
 
 
     /**
-     * 访客连接成功后设置的初始化缓存
+     * 访客连接成功后设置的初始化缓存(手动对话结束)
      * @param visitorId
      * @param webStyleId
      */
@@ -65,15 +100,56 @@ public class KfCacheRelUtils {
 
         //删除指定访客的kfsession
         String initKey = OneChatCacheKeyConstants.ImKeyGenerator.getInitVisitorSessionKey(visitorId,webStyleId);
-        redisService.deleteObject(initKey);
+        GuestIdentityMsgDto msgDto = SpringUtils.getBean(RedisService.class).getCacheObject(initKey);
 
-        //接触指定样式-客服规则下访客与客服的连接关系
-        ruleFfServingUtils.unbindVisitorFromKf(webStyleId,kfRuleId,userAccount,visitorId);
+        if(null != msgDto){
+            redisService.deleteObject(initKey);
+
+            //接触指定样式-客服规则下访客与客服的连接关系
+            ruleFfServingUtils.unbindVisitorFromKf(webStyleId,kfRuleId,userAccount,visitorId);
+
+
+
+            R<OneChatKfRule> data = remoteKfRuleService.findOneChatKfRule(Long.parseLong(kfRuleId), SecurityConstants.INNER);
+            if(null != data){
+                OneChatKfRule oneChatKfRule = data.getData();
+                if(null != oneChatKfRule){
+                    WsResponse response = WsResponse.fromText(JSONUtil.toJsonStr(GuestReplyMsgDto.buildGuestReplyMsgDto(OneChatKfVisitorMsg.builder()
+                            .kfVisitorId(Long.parseLong(msgDto.getKfVisitorId()))
+                            .fromObj(msgDto.getReceptObjId())
+                            .showAvatar(msgDto.getReceptObjavatar())
+                            .showName(msgDto.getReceptObjName())
+                            .toObj(msgDto.getVisitorId())
+                            .content(oneChatKfRule.getEndMsg())
+                            .kfRuleId(Long.parseLong(msgDto.getKfRuleId()))
+                            .msgType(OneChatMsgTypes.MSG_TYPE_TEXT)
+                            .sendTime(new Date())
+                            .build(),1,"endChat")), Constants.UTF8);
+                    String toUserId = new VisitorSessionKey(visitorId,webStyleId).toString();
+                    Tio.sendToUser(oneChatImStarter.getServerTioConfig(),toUserId, response);
+
+                }
+
+
+
+            }
+
+
+        }
+
+
+
+
+
+
+
+
+
     }
 
 
     /**
-     * 更新初始化连接数据
+     * 更新初始化连接数据（转人工）
      * @param visitorId
      * @param webStyleId
      */
@@ -83,9 +159,54 @@ public class KfCacheRelUtils {
 
         GuestIdentityMsgDto msgDto = SpringUtils.getBean(RedisService.class).getCacheObject(initKey);
         if(null != msgDto){
-            msgDto.setInitState(1);
-            msgDto.setVisitorMsg(null); //设置为空避免接入语重复
-            redisService.setCacheObject(initKey,msgDto, oneChatConfig.sessionTime.longValue() , TimeUnit.MINUTES);
+            R<OneChatKfRule> data = remoteKfRuleService
+                    .findOneChatKfRule(Long.parseLong(msgDto.getKfRuleId()), SecurityConstants.INNER);
+            OneChatKfRule oneChatKfRule = data.getData();
+
+            if(null != oneChatKfRule){
+
+                List<OneChatKfRuleScope> ruleScopeList = oneChatKfRule.getRuleScopeList();
+
+                OneChatKfRuleScope oneChatKfRuleScope = ruleScopeList.stream()
+                        .filter(item -> item.getUserAccount().equals(msgDto.getReceptObjId())).findAny().orElse(null);
+
+                if(null != oneChatKfRuleScope){
+                    msgDto.setInitState(OneChatVisitorSate.RECEIVE_STATE);
+                    msgDto.setReceptObjName(oneChatKfRuleScope.getNickName());
+                    msgDto.setReceptObjavatar(oneChatKfRuleScope.getAvatar());
+                    msgDto.setReceptObjId(oneChatKfRuleScope.getUserAccount());
+                    msgDto.setMsgTip(oneChatKfRule.getReceiveMsg());
+                    redisService.setCacheObject(initKey,msgDto, oneChatConfig.sessionTime.longValue() , TimeUnit.MINUTES);
+
+                    OneChatkfVisitor oneChatkfVisitor = remoteKfVisitorService.getOneChatkfVisitorById(Long.parseLong(msgDto.getKfVisitorId()), SecurityConstants.INNER).getData();
+
+                    if(oneChatkfVisitor != null){
+                        oneChatkfVisitor.setCurrentState(OneChatVisitorSate.getVisitorState(OneChatVisitorSate.RECEIVE_STATE));
+
+                        //更新访客状态
+                        remoteKfVisitorService.addOrUpdate(oneChatkfVisitor,SecurityConstants.INNER);
+
+                        WsResponse response = WsResponse.fromText(JSONUtil.toJsonStr(GuestReplyMsgDto.buildGuestReplyMsgDto(OneChatKfVisitorMsg.builder()
+                                .kfVisitorId(Long.parseLong(msgDto.getKfVisitorId()))
+                                .fromObj(msgDto.getReceptObjId())
+                                .showAvatar(oneChatKfRuleScope.getAvatar())
+                                .showName(oneChatKfRuleScope.getNickName())
+                                .toObj(msgDto.getVisitorId())
+                                .content(oneChatKfRule.getReceiveMsg())
+                                .kfRuleId(Long.parseLong(msgDto.getKfRuleId()))
+                                .msgType(OneChatMsgTypes.MSG_TYPE_TEXT)
+                                .sendTime(new Date())
+                                .build(),1,"chatting")), Constants.UTF8);
+                        String toUserId = new VisitorSessionKey(visitorId,webStyleId).toString();
+                        Tio.sendToUser(oneChatImStarter.getServerTioConfig(),toUserId, response);
+
+                    }
+
+
+
+                }
+
+            }
         }
 
     }
